preprocessing/requirements.txt
aioboto3==12.4.0
aiobotocore==2.12.3
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.10.1
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.16.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.11.0
apache-airflow-providers-http==4.13.0
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
apprise==1.9.3
argcomplete==3.5.0
asgi-lifespan==2.1.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
blinker==1.8.2
bokeh @ file:///tmp/abs_34854e1f-d7d3-4f22-85d9-1075588e4ecdga64o0qg/croots/recipe/bokeh_1658136654619/work
boto3==1.34.69
botocore==1.34.69
Brotli @ file:///croot/brotli-split_1736182456865/work
build==1.2.2.post1
cachelib==0.9.0
cachetools==6.1.0
certifi @ file:///croot/certifi_1750863846190/work/certifi
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1662587369221/work
cloudwatch==1.2.1
colorama==0.4.6
coloredlogs==14.0
colorlog==6.8.2
ConfigUpdater==3.2
connexion==2.14.2
coolname==2.2.0
cron-descriptor==1.4.5
cryptography==43.0.1
cssselect==1.3.0
cssutils==2.11.1
cytoolz @ file:///croot/cytoolz_1736803755750/work
dagster==1.8.8
dagster-cloud==1.8.8
dagster-cloud-cli==1.8.8
dagster-graphql==1.8.8
dagster-pipes==1.8.8
dagster-webserver==1.8.8
# Editable install with no version control (dagster_documents_orchestration==0.0.0)
-e /mnt/c/Users/karol/DataspellProjects/master-project/dagster-documents-orchestration
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1666724765694/work
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.14
dill==0.3.8
distributed @ file:///home/conda/feedstock_root/build_artifacts/distributed_1666295404135/work
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.13
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.3.1
github3.py==4.0.1
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
GPUtil==1.4.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.1.0
griffe==1.7.3
grpcio==1.66.1
grpcio-health-checking==1.62.3
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
HeapDict @ file:///Users/ktietz/demo/mc3/conda-bld/heapdict_1630598515714/work
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
idna @ file:///croot/idna_1714398848350/work
imagesize==1.4.1
importlib_metadata==8.4.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1741710844255/work
jinja2-humanize-extension==0.4.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.17
langchain-aws==0.1.4
langchain-community==0.3.26
langchain-core==0.2.43
langchain-openai==0.1.20
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lmdb @ file:///croot/python-lmdb_1736540601428/work
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
lockfile==0.12.2
lxml==5.4.0
lz4 @ file:///home/conda/feedstock_root/build_artifacts/lz4_1666828235631/work
Mako==1.3.5
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
methodtools==0.4.7
more-itertools==10.5.0
msgpack @ file:///home/conda/feedstock_root/build_artifacts/msgpack-python_1666755075942/work
multidict==6.1.0
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.37.1
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
orjson==3.10.18
packaging==24.1
paddleocr==3.0.2
paddlex==3.0.2
pandas==1.5.0
partd @ file:///croot/partd_1736803412080/work
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six @ file:///home/conda/feedstock_root/build_artifacts/pdfminer.six_1703854269349/work
pdfplumber @ file:///home/conda/feedstock_root/build_artifacts/pdfplumber_1735886669036/work
pendulum==3.1.0
pex==2.20.1
pillow==11.0.0
pip-tools==7.4.1
pluggy==1.5.0
plum-dispatch==1.7.4
prefect==3.4.6
prefect-dask==0.3.3
premailer==3.10.0
prettytable==3.16.0
prison==0.2.1
prometheus_client==0.22.1
prompt-toolkit==3.0.36
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow @ file:///croot/pyarrow_1707330824290/work/python
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome @ file:///croot/pycryptodome_1737039435377/work
pydantic==2.9.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.13.2
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1747560772891/work
PyPDF2 @ file:///croot/pypdf2_1733841476831/work
pypdfium2==4.30.1
pyproject_hooks==1.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytesseract==0.3.13
pytest==8.3.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nvd3==0.16.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2024.2
PyYAML==6.0.2
questionary==2.0.1
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3fs==2024.5.0
s3transfer==0.10.3
scikit-learn==1.7.0
scipy==1.15.3
setproctitle==1.3.3
shapely==2.1.1
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
spire.pdf==9.8.0
SQLAlchemy==2.0.41
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
starlette==0.46.2
structlog==24.4.0
tabulate==0.9.0
tblib @ file:///Users/ktietz/demo/mc3/conda-bld/tblib_1629402031467/work
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
time-machine==2.15.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
toposort==1.10
tornado @ file:///home/builder/ci_310/tornado_1640795392144/work
tqdm==4.67.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.5
upstash-ratelimit==1.1.0
upstash-redis==1.4.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1718978550903/work
uv==0.7.15
uvicorn==0.30.6
uvloop==0.20.0
Wand @ file:///home/conda/feedstock_root/build_artifacts/wand_1661205278277/work
watchdog==5.0.2
watchfiles==0.24.0
watchtower==3.3.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.11.1
zict @ file:///croot/zict_1695832839030/work
zipp==3.23.0
zstandard==0.23.0
